% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlacksburgVA.R
\docType{data}
\name{BlacksburgVA}
\alias{BlacksburgVA}
\title{Example meteorological training data for weather generator}
\format{
A data frame.
}
\source{
\href{https://www.ncdc.noaa.gov/cdo-web/datasets/GHCND/stations/GHCND:USC00440766/detail}{Blacksburg, VA NWS office}
}
\usage{
data(BlacksburgVA)
}
\description{
Weather data (precipitation, temperature, and season) measured at the NWS station
(GHCND:USC00440766) in Blacksburg, Virginia.
}
\examples{
data(BlacksburgVA)
}
\keyword{data}
