% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOCO_sims.R
\docType{data}
\name{BOCO_sims}
\alias{BOCO_sims}
\title{wxgenR simulation results for multiple meteorological stations in Boulder County, CO}
\format{
A list of dataframes.
}
\source{
\href{https://www.ncei.noaa.gov/products/land-based-station/global-historical-climatology-network-daily}{GHCNd}
}
\usage{
data(BOCO_sims)
}
\description{
Simulated precipitation and temperature at the following GHCN stations (data retrieved using the FedData R package):
"USC00050848" "USS0005J42S"
}
\examples{
data(BOCO_sims)
}
\keyword{data}
