% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowerSantaCruzRiverBasin.R
\docType{data}
\name{LowerSantaCruzRiverBasinAZ}
\alias{LowerSantaCruzRiverBasinAZ}
\title{Example meteorological training data for weather generator}
\format{
A data frame
}
\source{
\href{https://www.usbr.gov/lc/phoenix/programs/lscrbasin/LSCRBS_Hydroclimate_2021.pdf}{Hydroclimate Analysis - Lower Santa Cruz River Basin Study}
}
\usage{
data(LowerSantaCruzRiverBasinAZ)
}
\description{
Weather data (precipitation, temperature, and season) for the Lower Santa Cruz River Basin in Southern Arizona.
Dataset was developed for the Hydroclimate Analysis within Reclamation's Lower Santa Cruz River Basin Study.
}
\examples{
data(LowerSantaCruzRiverBasinAZ)
}
\keyword{data}
