% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationData.R
\docType{data}
\name{stationData}
\alias{stationData}
\title{Meteorological training data for weather generator from nine stations}
\format{
A data frame.
}
\source{
\href{https://www.ncei.noaa.gov/products/land-based-station/global-historical-climatology-network-daily}{Global Historical Climatology Network}
}
\usage{
data(stationData)
}
\description{
* Weather data (precipitation, temperature) from nine stations within the Global Historical Climatology Network (GHCN):
 + USC00162534	(DONALDSONVILLE 4 SW LA, US)
 + USC00050372	(ASPEN 1 SW CO, US)
 + USC00080236	(ARCHBOLD BIO STATION FL, US)
 + USC00440766	(BLACKSBURG NWS VA, US)
 + USW00014606	(BANGOR INTL. AIRPORT ME, US)
 + USW00094240	(QUILLAYUTE AIRPORT WA, US)
 + USC00346386	(NORMAN 3 SSE OK, US)
 + USC00028795	(TUCSON 17 NW AZ, US)
 + USC00472314	(EAGLE RIVER WI, US)
}
\examples{
data(stationData)
}
\keyword{data}
