% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wxgenR.R
\docType{package}
\name{wxgenR-package}
\alias{wxgenR-package}
\title{wxgenR: A Stochastic Weather Generator with Seasonality}
\description{
A weather generator to simulate precipitation and temperature for regions with seasonality. Users input training data containing precipitation, temperature, and seasonality (up to 26 seasons). Including weather season as a training variable allows users to explore the effects of potential changes in season duration as well as average start- and end-time dates due to phenomena like climate change. Data for training should be a single time series but can originate from station data, basin averages, grid cells, etc. Bearup, L., Gangopadhyay, S., & Mikkelson, K. (2021). "Hydroclimate Analysis Lower Santa Cruz River Basin Study (Technical Memorandum No ENV-2020-056)." Bureau of Reclamation. Gangopadhyay, S., Bearup, L. A., Verdin, A., Pruitt, T., Halper, E., & Shamir, E. (2019, December 1). "A collaborative stochastic weather generator for climate impacts assessment in the Lower Santa Cruz River Basin, Arizona." Fall Meeting 2019, American Geophysical Union. \url{https://ui.adsabs.harvard.edu/abs/2019AGUFMGC41G1267G}.
}
\author{
\strong{Maintainer}: David Woodson \email{dwoodson@usbr.gov}

Authors:
\itemize{
  \item David Woodson \email{dwoodson@usbr.gov}
  \item Subhrendu Gangopadhyay
  \item Lindsay Bearup
  \item Marketa McGuire
  \item Andrew Verdin
  \item Eylon Shamir
  \item Eve Halper
}

}
\keyword{internal}
