% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wxgenR.R
\name{wxgenR}
\alias{wxgenR}
\title{\code{wxgenR} package}
\description{
wxgenR: A weather generator with seasonality
}
\details{
This package provides functions for weather generation that incorporate seasonality.
}
